refactor: crete separated folders in routes
